Refactor : KafkaEvent 수정
