+++ concert/src/main/java/com/hhp/concert/Business/Domain/event/DataPlatformSendEvent.java
@@ -1,4 +1,4 @@
-package com.hhp.concert.Business.Domain;
+package com.hhp.concert.Business.Domain.event;
 
 
 public record DataPlatformSendEvent<T>(T data) {
