+++ concert/src/main/java/com/hhp/concert/Interfaces/consumer/OutboxEventListener.java
@@ -1,19 +1,22 @@
 package com.hhp.concert.Interfaces.consumer;
 
+import lombok.Getter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.stereotype.Component;
 
+
 @Component
-public class DataPlatFormEventListener {
+@Getter
+public class OutboxEventListener {
+    private static final Logger log = LoggerFactory.getLogger(OutboxEventListener.class);
 
-    private static final Logger log = LoggerFactory.getLogger(DataPlatFormEventListener.class);
-
-    @KafkaListener(topics = "concert-reserve-data", groupId = "${spring.kafka.consumer.group-id}")
-    public void consume(Object message) {
+    @KafkaListener(topics = "concert-reserve-data")
+    public void consume(String message) {
         log.info("consume");
         log.info("Consumed message: {}", message);
     }
+
 
 }
